Refactored dataset loaders Removed online dataset generator Refactored models H5 file generation now in line with prepare_dataset.m
