+++ rt_gene_model_training/pytorch/TestFastRTGENE.py
@@ -8,7 +8,7 @@ import torch
 from PIL import Image
 from torchvision import transforms
 
-from RTGENEModel_Mobilenetv2 import RTGENEModelMobileNetV2
+from RTGENEModels import RTGENEModelMobileNetV2
 from rt_gene.estimate_gaze_base import GazeEstimatorBase
 from rt_gene.extract_landmarks_method_base import LandmarkMethodBase
 from rt_gene.gaze_tools import get_phi_theta_from_euler, limit_yaw
@@ -16,17 +16,19 @@ from rt_gene.gaze_tools import get_phi_theta_from_euler, limit_yaw
 __script_path = os.path.dirname(os.path.realpath(__file__))
 
 _landmark_estimator = LandmarkMethodBase(device_id_facedetection="cuda:0",
-                                         checkpoint_path_face=os.path.abspath(os.path.join(__script_path, "../../rt_gene/model_nets/SFD/s3fd_facedetector.pth")),
-                                         checkpoint_path_landmark=os.path.abspath(os.path.join(__script_path, "../../rt_gene/model_nets/phase1_wpdc_vdc.pth.tar")),
+                                         checkpoint_path_face=os.path.abspath(
+                                             os.path.join(__script_path, "../../rt_gene/model_nets/SFD/s3fd_facedetector.pth")),
+                                         checkpoint_path_landmark=os.path.abspath(
+                                             os.path.join(__script_path, "../../rt_gene/model_nets/phase1_wpdc_vdc.pth.tar")),
                                          model_points_file=os.path.abspath(os.path.join(__script_path, "../../rt_gene/model_nets/face_model_68.txt")))
-_transform = transforms.Compose([transforms.CenterCrop(224),
+
+_transform = transforms.Compose([transforms.Resize((224, 224), Image.BICUBIC),
                                  transforms.ToTensor(),
-                                 transforms.Normalize(mean=[0.485, 0.456, 0.406],
-                                                      std=[0.229, 0.224, 0.225])])
+                                 transforms.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225])])
 
 device_id_gazeestimation = "cuda:0"
 _model = RTGENEModelMobileNetV2()
-_torch_load = torch.load(os.path.join(__script_path, "../model_nets/rt_gene_pytorch_checkpoints/_ckpt_epoch_43w.ckpt"))['state_dict']
+_torch_load = torch.load(os.path.join(__script_path, "../model_nets/rt_gene_pytorch_checkpoints/_ckpt_epoch_3.ckpt"))['state_dict']
 # for some reason, the state_dict appends _model infront of the states, remove it:
 _state_dict = {k[7:]: v for k, v in _torch_load.items()}
 _model.load_state_dict(_state_dict)
@@ -46,15 +48,13 @@ def extract_eye_image_patches(subject):
 
 
 try:
-    for file in _img_list:
-    #_cap = cv2.VideoCapture(2)
-    #while True:
-        frame = cv2.imread(file)
-        ret = True
-        # ret, frame = _cap.read()
+    _cap = cv2.VideoCapture(0)
+    while True:
+    # for file in _img_list:
+        # frame = cv2.imread(file)
+        # ret = True
+        ret, frame = _cap.read()
         if ret:
-            cv2.imshow("frame", frame)
-            stime = time.time()
             im_width, im_height = frame.shape[1], frame.shape[0]
             _dist_coefficients, _camera_matrix = np.zeros((1, 5)), np.array(
                 [[im_height, 0.0, im_width / 2.0], [0.0, im_height, im_height / 2.0], [0.0, 0.0, 1.0]])
@@ -83,19 +83,25 @@ try:
 
                 head_pose = get_phi_theta_from_euler(roll_pitch_yaw)
 
-                _transformed_left = _transform(Image.fromarray(subject.left_eye_color.astype('uint8'), 'RGB')).to(device_id_gazeestimation).unsqueeze(0)
-                _transformed_right = _transform(Image.fromarray(subject.right_eye_color.astype('uint8'), 'RGB')).to(device_id_gazeestimation).unsqueeze(0)
+                _left = Image.fromarray(subject.left_eye_color.astype('uint8'), 'RGB')
+                _right = Image.fromarray(subject.right_eye_color.astype('uint8'), 'RGB')
+
+                _transformed_left = _transform(_left).to(device_id_gazeestimation).unsqueeze(0)
+                _transformed_right = _transform(_right).to(device_id_gazeestimation).unsqueeze(0)
                 _head_pose = torch.from_numpy(np.array([*head_pose])).to(device_id_gazeestimation).unsqueeze(0).float()
 
+                stime = time.time()
                 gaze = _model(_transformed_left, _transformed_right, _head_pose)
-
                 gaze = gaze.detach().cpu().numpy().tolist()
+
+                print("Hz: {}".format(1.0 / (time.time() - stime)))
                 l_gaze_img = GazeEstimatorBase.visualize_eye_result(subject.left_eye_color, gaze[0])
                 r_gaze_img = GazeEstimatorBase.visualize_eye_result(subject.right_eye_color, gaze[0])
                 s_gaze_img = np.concatenate((r_gaze_img, l_gaze_img), axis=1)
 
-                cv2.imshow("patches", s_gaze_img)
-                cv2.waitKey(0)
+                cv2.imshow("face", frame)
+                cv2.imshow("patches".format(*gaze[0][::-1]), s_gaze_img)
+                cv2.waitKey(1)
 
 except KeyboardInterrupt:
     cv2.destroyAllWindows()
