+++ rt_gene_model_training/pytorch/TrainRTGENELightening.py
@@ -5,20 +5,17 @@ import h5py
 import numpy as np
 import pytorch_lightning as pl
 import torch
+from PIL import Image, ImageFilter
 from pytorch_lightning.callbacks import ModelCheckpoint
 from torch.utils.data import DataLoader
+from torchvision.transforms import transforms
 
 from GazeAngleAccuracy import GazeAngleAccuracy
-from RTGENEH5Dataset import RTGENEDataset
-from RTGENEModel_Mobilenetv2 import RTGENEModelMobileNetV2
-from RTGENEModel_Resnet import RTGENEModelResnet18, RTGENEModelResnet50
-from RTGENEModel_VGG16 import RTGENEModelVGG
+from RTGENEDataset import RTGENEH5Dataset
+from RTGENEModels import RTGENEModelMobileNetV2, RTGENEModelResnet18, RTGENEModelResnet50, RTGENEModelVGG
 
 
 class TrainRTGENE(pl.LightningModule):
-
-    def tng_dataloader(self):
-        pass
 
     def __init__(self, hparams):
         super(TrainRTGENE, self).__init__()
@@ -55,11 +52,10 @@ class TrainRTGENE(pl.LightningModule):
         return {'val_loss': loss, 'angle_acc': angle_acc}
 
     def validation_end(self, outputs):
-        avg_loss = torch.stack([x['val_loss'] for x in outputs]).mean()
-        avg_angle = np.array([x['angle_acc'] for x in outputs])
-        avg_angle = np.mean(avg_angle)
-        tensorboard_logs = {'val_loss': avg_loss, 'val_angle_acc': avg_angle}
-        return {'val_loss': avg_loss, 'log': tensorboard_logs}
+        _losses = torch.stack([x['val_loss'] for x in outputs])
+        _angles = np.array([x['angle_acc'] for x in outputs])
+        tensorboard_logs = {'val_loss': _losses.mean(), 'val_angle_acc': np.mean(_angles), 'val_angle_std': np.std(_angles)}
+        return {'val_loss': _losses.mean(), 'log': tensorboard_logs}
 
     def configure_optimizers(self):
         _params_to_update = []
@@ -77,13 +73,20 @@ class TrainRTGENE(pl.LightningModule):
 
     @pl.data_loader
     def train_dataloader(self):
-        _data_train = RTGENEDataset(h5_file=h5py.File(self.hparams.hdf5_file, mode="r"), subject_list=[0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14])
-        return DataLoader(_data_train, batch_size=self.hparams.batch_size, shuffle=True)
+        _train_transforms = transforms.Compose([transforms.RandomResizedCrop(size=(224, 224), scale=(0.85, 1.0)),
+                                                transforms.Resize((224, 224), Image.BICUBIC),
+                                                transforms.RandomGrayscale(p=0.08),
+                                                lambda x: x if np.random.random_sample() > 0.08 else x.filter(ImageFilter.GaussianBlur(radius=2)),
+                                                lambda x: x if np.random.random_sample() > 0.0064 else x.filter(ImageFilter.GaussianBlur(radius=2)),
+                                                transforms.ToTensor(),
+                                                transforms.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225])])
+        _data_train = RTGENEH5Dataset(h5_file=h5py.File(self.hparams.hdf5_file, mode="r"), subject_list=[0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14])
+        return DataLoader(_data_train, batch_size=self.hparams.batch_size, shuffle=True, num_workers=4)
 
     @pl.data_loader
     def val_dataloader(self):
-        _data_validate = RTGENEDataset(h5_file=h5py.File(self.hparams.hdf5_file, mode="r"), subject_list=[15, 16])
-        return DataLoader(_data_validate, batch_size=self.hparams.batch_size, shuffle=True)
+        _data_validate = RTGENEH5Dataset(h5_file=h5py.File(self.hparams.hdf5_file, mode="r"), subject_list=[15, 16])
+        return DataLoader(_data_validate, batch_size=self.hparams.batch_size, shuffle=True, num_workers=4)
 
 
 if __name__ == "__main__":
@@ -95,7 +98,7 @@ if __name__ == "__main__":
 
     # gpu args
     _root_parser.add_argument('--gpus', type=int, default=1, help='how many gpus')
-    _root_parser.add_argument('--learning_rate', default=0.001, type=float)
+    _root_parser.add_argument('--learning_rate', default=0.00075, type=float)
     _root_parser.add_argument('--model_base', choices=["vgg", "mobilenet", "resnet18", "resnet50"], default="vgg")
     _root_parser.add_argument('--hdf5_file', default=os.path.abspath("/home/ahmed/Documents/RT_GENE/dataset.hdf5"), type=str)
     _root_parser.add_argument('--save_dir', default=os.path.abspath(os.path.join(root_dir, '..', 'model_nets', 'rt_gene_pytorch_checkpoints')))
@@ -106,7 +109,7 @@ if __name__ == "__main__":
 
     _model = TrainRTGENE(hparams=_hyperparams)
 
-    checkpoint_callback = ModelCheckpoint(filepath=_hyperparams.save_dir, monitor='val_loss', mode='min', verbose=True, save_top_k=-1)  # save all the models
+    checkpoint_callback = ModelCheckpoint(filepath=_hyperparams.save_dir, monitor='val_loss', mode='min', verbose=True, save_top_k=3)  # save the best models
 
     # earlystopping_callback = EarlyStopping(monitor="val_loss", patience=10, mode="min", verbose=True)
